all but the initialization of symboltable is done with the lexer

+++ src/lexer/token.rs
@@ -13,7 +13,7 @@ pub enum TokenType {
     Is,
 
     // Types
-    Integer(i64),
+    Integer,
     Boolean,
 
     // Literals
